+++ lib/quick_button.dart
@@ -1,19 +1,23 @@
+/// Import Quick Button Library.
 library quick_button;
 
 // ignore_for_file: prefer_const_constructors, prefer_const_literals_to_create_immutables, sort_child_properties_last, must_be_immutable
 
 import 'package:flutter/material.dart';
 
+/// Create a stateful widget class
 class QuickButton extends StatefulWidget {
-  String labelText;
+  String? labelText;
   Color? labelColor;
   double? labelSize;
   FontWeight? labelWeight;
   double? labelSpacing;
+  IconData? icon;
   IconData? prefixIcon;
   IconData? suffixIcon;
   double? iconSize;
-  double? buttonSize;
+  double? buttonWidth;
+  double? buttonHeight;
   double? borderRadius;
   Color? borderColor;
   double? borderSize;
@@ -22,20 +26,21 @@ class QuickButton extends StatefulWidget {
   Color? backgroundHoverIn;
   Color? labelHoverOut;
   Color? backgroundHoverOut;
-  String? imageIcon;
   final Function onPressed;
 
   QuickButton(
       {Key? key,
-      required this.labelText,
+      this.labelText,
       this.labelColor,
       this.labelSize,
       this.labelWeight,
       this.labelSpacing,
+      this.icon,
       this.prefixIcon,
       this.suffixIcon,
       this.iconSize,
-      this.buttonSize,
+      this.buttonWidth,
+      this.buttonHeight,
       this.borderRadius,
       this.borderColor,
       this.borderSize,
@@ -44,7 +49,6 @@ class QuickButton extends StatefulWidget {
       this.backgroundHoverIn,
       this.labelHoverOut,
       this.backgroundHoverOut,
-      this.imageIcon,
       required this.onPressed})
       : super(key: key);
 
@@ -57,6 +61,7 @@ class _QuickButtonState extends State<QuickButton> {
   Widget build(BuildContext context) {
     double width = MediaQuery.of(context).size.width;
 
+    /// Button Font style.
     final buttonFont = TextStyle(
         fontSize: widget.labelSize ?? 13,
         color: widget.labelColor ?? Colors.white,
@@ -76,88 +81,132 @@ class _QuickButtonState extends State<QuickButton> {
           widget.backgroundColor = widget.backgroundHoverOut;
         });
       },
-      child: SizedBox(
-        width: widget.buttonSize ?? 200.0,
-        child: TextButton(
-          child: Container(
-            padding: width >= 800 ? EdgeInsets.all(8) : EdgeInsets.all(6),
-            child: FittedBox(
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  if (widget.suffixIcon == null &&
-                      widget.prefixIcon != null) ...{
-                    const SizedBox(
-                      width: 18,
-                    ),
-                    Icon(
-                      widget.prefixIcon,
-                      color: widget.labelColor,
-                      size: widget.iconSize ?? 15,
-                    ),
-                    const SizedBox(
-                      width: 7,
-                    ),
-                    Text(
-                      widget.labelText,
-                      style: buttonFont,
-                    ),
-                    Icon(
-                      widget.suffixIcon,
-                      color: widget.labelColor,
-                      size: widget.iconSize ?? 15,
-                    ),
-                  } else if (widget.prefixIcon == null &&
-                      widget.suffixIcon != null) ...{
-                    Icon(
-                      widget.prefixIcon,
-                      color: widget.labelColor,
-                      size: widget.iconSize ?? 15,
-                    ),
-                    Text(
-                      widget.labelText,
-                      style: buttonFont,
-                    ),
-                    const SizedBox(
-                      width: 7,
-                    ),
-                    Icon(
-                      widget.suffixIcon,
-                      color: widget.labelColor,
-                      size: widget.iconSize ?? 15,
-                    ),
-                    const SizedBox(
-                      width: 18,
-                    ),
-                  } else if (widget.prefixIcon == null ||
-                      widget.suffixIcon == null) ...{
-                    Text(
-                      widget.labelText,
-                      style: buttonFont,
-                    ),
-                  }
-                ],
-              ),
-            ),
-          ),
-          onPressed: () {
+      child: InkWell(
+          key: const Key('onTapButton'),
+          onTap: () {
             widget.onPressed();
           },
-          style: TextButton.styleFrom(
-            backgroundColor:
-                widget.backgroundColor ?? Color.fromARGB(255, 31, 31, 31),
-            shape: RoundedRectangleBorder(
-              side: BorderSide(
-                  color: widget.borderColor ?? Colors.transparent,
-                  width: widget.borderSize ?? 1),
-              borderRadius: BorderRadius.circular(
-                widget.borderRadius ?? 0.0,
-              ),
+          child: Container(
+            key: const Key('buttonDecoration'),
+            width: widget.buttonWidth ?? 150.0,
+            height: widget.buttonHeight ?? 35.0,
+            padding: width >= 800
+                ? widget.labelText == null
+                    ? const EdgeInsets.all(4)
+                    : const EdgeInsets.all(6)
+                : const EdgeInsets.all(5),
+            alignment: Alignment.center,
+
+            /// Button decoration.
+            decoration: BoxDecoration(
+                color: widget.backgroundColor ??
+                    const Color.fromARGB(255, 31, 31, 31),
+                borderRadius: BorderRadius.circular(
+                  widget.borderRadius ?? 0.0,
+                ),
+                border: Border.all(
+                    color: widget.borderColor ?? Colors.transparent,
+                    width: widget.borderSize ?? 1)),
+            child: FittedBox(
+              child: Row(children: [
+                /// Button without prefixIcon.
+                if (widget.suffixIcon == null && widget.prefixIcon != null) ...{
+                  const SizedBox(
+                    width: 18,
+                  ),
+                  Icon(
+                    widget.prefixIcon,
+                    color: widget.labelColor,
+                    size: widget.iconSize ?? 15,
+                  ),
+                  const SizedBox(
+                    width: 7,
+                  ),
+                  widget.labelText == null
+                      ? Icon(
+                          widget.icon,
+                          size: widget.iconSize,
+                          color: widget.labelColor,
+                        )
+                      : Text(
+                          key: const Key('labelText'),
+                          widget.labelText.toString(),
+                          style: buttonFont,
+                        ),
+                  Icon(
+                    widget.suffixIcon,
+                    color: widget.labelColor,
+                    size: widget.iconSize ?? 15,
+                  ),
+                }
+
+                /// Button with suffixIcon.
+                else if (widget.prefixIcon == null &&
+                    widget.suffixIcon != null) ...{
+                  Icon(
+                    widget.prefixIcon,
+                    color: widget.labelColor,
+                    size: widget.iconSize ?? 15,
+                  ),
+                  widget.labelText == null
+                      ? Icon(
+                          widget.icon,
+                          size: widget.iconSize,
+                          color: widget.labelColor,
+                        )
+                      : Text(
+                          key: const Key('labelText'),
+                          widget.labelText.toString(),
+                          style: buttonFont,
+                        ),
+                  const SizedBox(
+                    width: 7,
+                  ),
+                  Icon(
+                    widget.suffixIcon,
+                    color: widget.labelColor,
+                    size: widget.iconSize ?? 15,
+                  ),
+                  const SizedBox(
+                    width: 18,
+                  ),
+                }
+
+                /// Button with labelText only.
+                else if (widget.prefixIcon == null ||
+                    widget.suffixIcon == null) ...{
+                  widget.labelText == null
+                      ? Icon(
+                          widget.icon,
+                          size: widget.iconSize,
+                          color: widget.labelColor,
+                        )
+                      : Text(
+                          key: const Key('labelText'),
+                          widget.labelText.toString(),
+                          style: buttonFont,
+                        ),
+                }
+
+                /// Button with Icon only.
+                else if (widget.prefixIcon == null ||
+                    widget.suffixIcon == null ||
+                    widget.labelText == null) ...{
+                  widget.labelText == null
+                      ? Icon(
+                          widget.icon,
+                          size: widget.iconSize,
+                          color: widget.labelColor,
+                        )
+                      : Text(
+                          key: const Key('labelText'),
+                          widget.labelText.toString(),
+                          style: buttonFont,
+                        ),
+                }
+              ]),
             ),
-          ),
-        ),
-      ),
+          )),
     );
   }
 }