+++ CHANGELOG.md
@@ -2,4 +2,12 @@
 
 * First release.
 
+## 0.1.0 - 14/08/2023
+
+* Depreciated **`buttonSize`** property.
+* Added **`icon`**, **`buttonWidth`** & **`buttonHeight`** properties.
+* Added Documentation.
+* Updated ReadME.md.
+
+